InMemoryMealRepository implemented

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -3,11 +3,13 @@ package ru.javawebinar.topjava.repository.inmemory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.util.Collection;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepository implements MealRepository {
     private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -25,22 +27,32 @@ public class InMemoryMealRepository implements MealRepository {
             return meal;
         }
         // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> {
+            if (oldMeal.getUserId() == SecurityUtil.authUserId()) {
+                return meal;
+            } else return oldMeal;
+        });
     }
 
     @Override
     public boolean delete(int id) {
-        return repository.remove(id) != null;
+        Meal meal = repository.get(id);
+        if (meal != null && meal.getUserId() == SecurityUtil.authUserId()) {
+            return repository.remove(id) != null;
+        }
+        return false;
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        Meal meal = repository.get(id);
+        return meal.getUserId() == SecurityUtil.authUserId() ? meal : null;
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        return repository.values().stream()
+                .filter(meal -> meal.getUserId() == SecurityUtil.authUserId()).sorted().collect(Collectors.toList());
     }
 }
